state_machine: add support for nack packets
Please note that this is just a temporary solution for windows
systems that have weird tcp behaviors which would cause tcp packet
corruption such that gdb would not recognize the packet then send
back a nack packet.

There should be a better solution while this one just record the last
packet in the struct and resend them when nack is issued.

Signed-off-by: Walter Ji <walter.ji@cipunited.com>