XPacket: compatibility for mips-mti-gdb
This commit is only here because mips-mti-gdb would send
sign extended addressed in XPackets, this causes gdbstub to
send a TargetMismatch error and panics.

The commit fixed this by using 64 bit address space as default,
then checks the max possible address of current target, and use
the size of bytes it occupies to truncate the address sent by gdb
and finally parse it with from_be_bytes.

Please note that this addition could cause weird behaviors with
other gdbs.

Signed-off-by: Walter Ji <walter.ji@cipunited.com>